Make a base for Pad entity and finish DetailedLocation
